editors: only extend line width to 119 for C and XML files
For all other files leave the line width at 79 as before. This is a good idea
since we generally don't want text files such as catalog files, unit files or
README/NEWS files to be line-broken at 119 since they are regularly browsed on
text terminals.

While we are at it, also add a couple of comments to the various files.

(Note that .editorconfig doesn't carry line-width information, simply because
the specification doesn't know the concept.)