cgroup: remove support for net_cls and net_prio controllers
Support for net_cls and net_prio controllers, along with the NetClass=
configuration directive, has been added in v227 in preparation for a
per-unit packet filter mechanism. However, it turns out the kernel people
have decided to deprecate that controller in v2. Tejun provides a
comprehensive justification for this in his commit, which has landed during
the merge window for kernel v4.5:

  https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=bd1060a1d671

As we're aiming for full support for the v2 cgroup hierarchy, we can no
longer support this feature. Userspace tool such as nftables are moving over
to setting rules that are specific to the full cgroup path of a task, which
obsoletes these controllers anyway.

This commit removes all support for the net_cls and net_prio controllers,
but keeps the NetClass= directive around for legacy compatibility reasons.