Allow DNSSEC mode of transactions to be specified
Before, the DNSSEC mode was specified for specific scopes based on
settings internal to the daemon. Allow the mode to be specified
by the client, as flags to ResolveHostname. Following flags are allowed:
- 0: the same as before, use server defaults
- SD_RESOLVED_AUTHENTICATED: force DNSSEC
- SD_RESOLVED_AUTHENTICATED|SD_RESOLVED_ALLOW_DOWNGRADE: try DNSSEC
- SD_RESOLVED_DONT_AUTHENTICATE: disable DNSSEC even if enabled in daemon

The mode can be specified with --dnssec[=yes|no|allow-downgrade] option
to systemd-resolve.

If is possible to receive an unauthenticated answer even with DNSSEC was
forced, if the domain is insecure (provably not signed by DNSSEC). It
might be useful to have a mode where insecure answers are not allowed,
but this patch does not provide that.