Allow getty@.service to be used on systems without virtual consoles
Instead of checking for the existence of /dev/tty0, check for the actual
TTY this service in instantiated for. This allows the getty service to
be reused for other console types without modifications.

In particular, this allows it to be used for the PTYs LXC allocate to
emulate TTYs, but it should also be useful for other systems that don't
use virtual consoles.