rules: remove SYSTEMD_READY=0 on DM_UDEV_DISABLE_OTHER_RULES_FLAG=1
The "SYSTEMD_READY=0" will cause automatic unmount of mountpoint that is
on top of such DM device. It's even more problematic if this is used
with multipath which sets DM_UDEV_DISABLE_OTHER_RULES_FLAG in case we
have a CHANGE event that comes after DM multipath device reload when one
of the paths is down or up.

See https://bugzilla.redhat.com/show_bug.cgi?id=1312011