locale: do not always assume UTF-8 support is available
In the absence of accurate locale information (via the LANG, LC_ALL,
LC_CTYPE, etc. environment variables), it is not safe for interactive
tools to assume UTF-8 support is available.

In particular, with missing locale information "less" (used as a systemd
pager) will misrender UTF-8 content, e.g.:

  $ ( unset LANG LC_ALL LC_CTYPE; systemctl status )
  <E2><97><8F> hostname.example.com
      State: running
       Jobs: 0 queued
     Failed: 0 units
      Since: Mon 2016-02-08 09:38:49 AEDT; 3 weeks 1 days ago
     CGroup: /
             <E2><94><9C><E2><94><80>1 /usr/lib/systemd/systemd --system --deserialize 22
             <E2><94><9C><E2><94><80>system.slice
             ...

Other pagers are likely to behave similarly.

For systemd-vconsole-setup, on the other hand, it *is* safe to assume
UTF-8 support is available if the locale information in the environment
is missing.