Add sleep for non-root calls of `systemd-notify`
Fixes #2379

Without `CAP_SYS_ADMIN` it's not possible to fake user credentials in
`sd_pid_notify` call, and in this case everything depends on timing -
whether systemd will be able to get the cgroup of sending process before
it exists or not.

Adding `sleep(3)` for non-root users will at least make `systemd-notify`
useful for them.