nspawn: overlay with lower from image
This makes it possible to overlay a root filesystem with
a writeable directory outside the container.

It changed the "two-arg special behavior", but this might
be ok, since these are fresh and afaik not very much used
features yet, and probably will not break "userspace".

"--overlay=/etc:/tmp/writable" will with this patch mount
/etc from inside disk-image as "lower"/underlay with
/tmp/writable from host as "upper"/overlay and place the
merged filesystem on /etc inside container.

translated to cmdline, this becomes: (split for readability)

mount -t overlay overlay \
      -o lowerdir=/tmp/nspawn-root-RND/etc,\
         upperdir=/tmp/writable,\
         workdir=/tmp/.#overlay7150c3069b11c61 \
      /tmp/nspawn-root-RND/etc