nspawn: populate /etc/machine-id if it is empty
If journal linking is explicitly requested but container image doesn't
have /etc/machine-id populated nspawn refuses to spawn such
container.

However, from user point of view it is reasonable to expect that container
image distributed w/ empty /etc/machine-id will be started and container
manager takes care for assignment of ID when container starts up.

Let's do just that and setup machine-id in cases we need it to link the
journals but file present in the image is empty.