rules: set SYSTEMD_READY=0 on DM_UDEV_DISABLE_OTHER_RULES_FLAG=1 only with ADD event
The "SYSTEMD_READY=0" will cause automatic unmount of mountpoint that is on top
of such a DM device if this is used with multipath (which sets
DM_UDEV_DISABLE_OTHER_RULES_FLAG). Prevent a CHANGE event that comes after DM
multipath device reload when one of the paths is down or up from marking the
device as not ready again.

https://bugzilla.redhat.com/show_bug.cgi?id=1312011
https://launchpad.net/bugs/1565969