[RFC] nspawn: defer sd_notify READY=1 until default.target
This patch is not ready yet. I am filing this pull request to add
questions and to keep that work somewhere: I will probably not continue
the work this week but after.

I am not sure when to enable the feature: when systemd-nspawn is not
started with --boot, we definitely don't want to defer READY=1. But even
with --boot, we might not want to defer READY=1 if the container has an
older version of systemd or something else. I don't know if we want to
have a new parameter on systemd-nspawn to enable the feature. I added
the environment variable SYSTEMD_NSPAWN_DEFER_NOTIFY_READY=1 for now.

systemd-nspawn listens on a file socket and systemd in the container
sends messages through the socket. I would like the socket to be
/run/systemd/nspawn/notify because that directory already has "incoming"
for communicating bind-mounts between the host and the container. But I
didn't manage to get that working because /run is a tmpfs, not shared
between the two mount namespaces. So for now, I just use /notify.

I am not sure where to hook the code when default.target is reached. For
now, I just added it somewhere in exec_spawn() for testing.

I am testing this with the following commands:

sudo SYSTEMD_NSPAWN_DEFER_NOTIFY_READY=1 $HOME/git/systemd/systemd-nspawn -D $HOME/distro-trees/fedora-22/ -b
sudo systemd-run --service-type=notify sh -c 'SYSTEMD_NSPAWN_DEFER_NOTIFY_READY=1 /home/alban/git/systemd/systemd-nspawn -D /home/alban/distro-trees/fedora-22/ -b'

https://github.com/systemd/systemd/issues/1369