networkd: fix #3046
IPv6 protocol requires a minimum MTU of 1280 bytes on the interface
hence this is not getting set.
This fixes 3046.

Introduce helper link_ipv6_enabled() to figure out whether IPV6 is
enabled.
Introduce network_has_static_ipv6_addresses() to find out if any static
ipv6 address configured.
If IPv6 is not configured on any interface that is SLAAC,
DHCPv6 and static IPv6 addresses not configured, then IPv6 will be
automatically disabled for that interface that is we write
"1" to /proc/sys/net/ipv6/conf//disable_ipv6.