sysctl.d: the conf file dirs may contain an additional "$(uname -r)" directory with conf files
Some distros makes possible to install several kernels with different
flavors. In that case a specific kernel is selected and started by the
bootloader.

Some kernel setting values might depend on the flavor of the kernel
which is booted by the system.

This patch adds additional directories in the list of conf directories
considered by systemd-sysctl. Those new dirs are located in the
default ones and named after the release of the running kernel.

This eases the selection of some kernel variables according to the
release of the running kernel.